backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose
pytesseract
pdf2image
Pillow
python-dotenv
passlib[bcrypt]
motor
